+++ roboplan_ros_cpp/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>roboplan_roscpp</name>
+  <name>roboplan_ros_cpp</name>
   <version>0.0.0</version>
   <description>ROS 2 C++ bindings for the roboplan motion planning library.</description>
   <maintainer email="sebas.a.castro@gmail.com">Sebastian Castro</maintainer>
