+++ roboplan_ros_py/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>roboplan_rospy</name>
+  <name>roboplan_ros_py</name>
   <version>0.0.0</version>
   <description>ROS 2 Python bindings for the roboplan motion planning library.</description>
   <maintainer email="sebas.a.castro@gmail.com">Sebastian Castro</maintainer>
@@ -11,6 +11,7 @@
   <depend>rclpy</depend>
   <depend>trajectory_msgs</depend>
   <depend>builtin_interfaces</depend>
+  <depend>roboplan_bindings</depend>
 
   <test_depend>python3-pytest</test_depend>
