+++ roboplan_ros_py/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/roboplan_ros_py
+[install]
+install_scripts=$base/lib/roboplan_ros_py
