--- roboplan_rospy/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/roboplan_rospy
-[install]
-install_scripts=$base/lib/roboplan_rospy
